update variable tests
