+++ packages/plugin-e2e/tests/as-admin-user/datasource/variables/variableInterpolation.spec.ts
@@ -1,49 +1,15 @@
 import { test, expect, PanelEditPage } from '../../../../src';
 
-test('variable interpolation', async ({
-  readProvisionedDashboard,
-  request,
-  page,
-  selectors,
-  grafanaVersion,
-}, testInfo) => {
-  const dashboard = await readProvisionedDashboard({ fileName: 'redshift.json' });
-  const panelEditPage = new PanelEditPage(
-    { request, page, selectors, grafanaVersion, testInfo },
-    {
-      id: '5',
-      dashboard,
-    }
-  );
+test('variable interpolation', async ({ readProvisionedDashboard, gotoDashboardPage }) => {
+  const dashboard = await readProvisionedDashboard({ fileName: 'testdatasource.json' });
+  const dashboardPage = await gotoDashboardPage(dashboard);
+  const panelEditPage = await dashboardPage.addPanel();
+  await panelEditPage.datasource.set('test-datasource');
+  const editorRow = await panelEditPage.getQueryEditorRow('A');
+  await editorRow.getByRole('textbox', { name: 'Query Text' }).fill('$var1');
   const queryReq = panelEditPage.waitForQueryDataRequest((request) =>
-    (request.postData() ?? '').includes(
-      `"rawSQL":"select * from long_format_example where environment in ('staging') limit 100"`
-    )
+    (request.postData() ?? '').includes(`"queryText":"A"`)
   );
-  await panelEditPage.goto();
-  await expect(await queryReq).toBeTruthy();
-});
-
-test('variable interpolation (navigate to panel from dashboard)', async ({
-  readProvisionedDashboard,
-  request,
-  page,
-  selectors,
-  grafanaVersion,
-}, testInfo) => {
-  const dashboard = await readProvisionedDashboard({ fileName: 'redshift.json' });
-  const panelEditPage = new PanelEditPage(
-    { request, page, selectors, grafanaVersion, testInfo },
-    {
-      id: '5',
-      dashboard,
-    }
-  );
-  const queryReq = panelEditPage.waitForQueryDataRequest((request) =>
-    (request.postData() ?? '').includes(
-      `"rawSQL":"select * from long_format_example where environment in ('staging') limit 100"`
-    )
-  );
-  await panelEditPage.goto();
+  await panelEditPage.refreshPanel();
   await expect(await queryReq).toBeTruthy();
 });
